// src/model/UserModel.ts
export class UserModel {
    id?: number; //Math.floor(Math.random() * (10000000)) + 1;
    firstName?: string;
    lastName?: string;
    emailAddress?: string;
    phoneNumber?: string;
    addressStreet?: string;
    addressNumber?: string;
    addressCity?: string;
    addressZipCode?: string;
}